Add shared DirectAccessIndexInput test utility
